+++ src/Components/Form/ArrayTextInput/ArrayTextInput.scss
@@ -1,4 +1,4 @@
-@import "../../colors";
+@import "../../../colors";
 
 .array-input-wrapper {
   display: flex;
