+++ src/Components/Form/Select/Select.scss
@@ -1,4 +1,4 @@
-@import "../../colors";
+@import "../../../colors";
 
 .select-input-container {
   width: 100%;
