+++ src/Components/Form/Switch/Switch.scss
@@ -1,4 +1,4 @@
-@import "../../colors";
+@import "../../../colors";
 
 .switch-wrapper {
   display: flex;
