+++ src/Components/Form/TextInput/TextInput.scss
@@ -1,4 +1,4 @@
-@import "../../colors";
+@import "../../../colors";
 
 .text-input-wrapper {
   position: relative;
