+++ src/Components/PackageForm/PackageForm.jsx
@@ -2,8 +2,8 @@ import React, { useCallback, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 
 import Button from "../Button/Button.jsx";
-import Select from "../Select/Select.jsx";
-import TextInput from "../TextInput/TextInput.jsx";
+import Select from "../Form/Select/Select.jsx";
+import TextInput from "../Form/TextInput/TextInput.jsx";
 
 import { createPackage } from "../../utils/api.js";
 import { languages } from "../../utils/constants.js";
